rename Aux->Auxillary to avoid Windows name issue
