+++ src/components/Burger/OrderSummary/OrderSummary.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Aux from '../../../hoc/Aux/Aux';
+import Aux from '../../../hoc/Auxillary/Auxillary';
 import Button from '../../UI/Button/Button';
 
 // class used to lifecycle hooks can be used
